+++ apps/web/vite.config.ts
@@ -8,6 +8,7 @@ export default defineConfig({
   resolve: {
     alias: {
       "@": path.resolve(__dirname, "./src"),
+      "@repo/ui/*": "../../packages/ui/src/*"
     },
   },
   server: {
