--- packages/ui/.eslintrc.cjs
@@ -1,5 +1,0 @@
-/** @type {import("eslint").Linter.Config} */
-module.exports = {
-  root: true,
-  extends: ["@repo/eslint-config/index.js"],
-};
