--- packages/ui/index.ts
@@ -1,4 +1,0 @@
-// components
-export * from "./src/components";
-// lib
-export * from "./src/lib";
