+++ packages/ui/src/components/label.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import * as LabelPrimitive from "@radix-ui/react-label"
 
-import { cn } from "../../lib/utils"
+import { cn } from "@repo/ui/lib/utils"
 
 function Label({
   className,
