+++ packages/ui/src/components/paragraph.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Slot } from "@radix-ui/react-slot";
 import { cva, type VariantProps } from "class-variance-authority";
-import { cn } from "../../lib/utils";
+import { cn } from "@repo/ui/lib/utils";
 
 const paragraphVariants = cva("leading-7 [&:not(:first-child)]:mt-6", {
   variants: {
