+++ packages/ui/src/components/toggle-group.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import * as ToggleGroupPrimitive from "@radix-ui/react-toggle-group"
 import { type VariantProps } from "class-variance-authority"
 
-import { cn } from "../../lib/utils"
+import { cn } from "@repo/ui/lib/utils"
 import { toggleVariants } from "./toggle"
 
 const ToggleGroupContext = React.createContext<
