+++ packages/ui/tsconfig.json
@@ -4,7 +4,7 @@
   "compilerOptions": {
     "jsx": "react-jsx",
     "baseUrl": ".",
-    "paths": { "@ui/*": ["./src/*"] },
+    "paths": { "@repo/ui/*": ["./src/*"] },
     "moduleResolution": "bundler"
   }
 }
